+++ game/client/cstrike/c_cs_player.h
@@ -249,7 +249,7 @@ public:
 	CNetworkVar( int, m_iProgressBarDuration );
 
 	// When the progress bar should start.
-	CNetworkVar( float, m_flProgressBarStartTime );
+	CNetworkVar( double, m_flProgressBarStartTime );
 
 	CNetworkVar( float, m_flStamina );
 	CNetworkVar( int, m_iDirection );	// The current lateral kicking direction; 1 = right,  0 = left
