+++ game/client/ragdoll.cpp
@@ -698,7 +698,7 @@ public:
 		m_vecOffset.Init();
 	}
 	DECLARE_CLIENTCLASS();
-	bool SetupBones( matrix3x4_t *pBoneToWorldOut, int nMaxBones, int boneMask, float currentTime )
+	bool SetupBones( matrix3x4_t *pBoneToWorldOut, int nMaxBones, int boneMask, double currentTime )
 	{
 		if ( GetMoveParent() )
 		{
