+++ game/server/playerlocaldata.h
@@ -50,10 +50,10 @@ public:
 	// In process of duck-jumping
 	CNetworkVar( bool, m_bInDuckJump );
 	// During ducking process, amount of time before full duc
-	CNetworkVar( float, m_flDucktime );
-	CNetworkVar( float, m_flDuckJumpTime );
+	CNetworkVar( double, m_flDucktime );
+	CNetworkVar( double, m_flDuckJumpTime );
 	// Jump time, time to auto unduck (since we auto crouch jump now).
-	CNetworkVar( float, m_flJumpTime );
+	CNetworkVar( double, m_flJumpTime );
 	// Step sound side flip/flip
 	int m_nStepside;;
 	// Velocity at time when we hit ground
